Teacher-Training/NEH/requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0

# Machine Learning
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress and Logging
tqdm>=4.65.0
tensorboard>=2.13.0

# Note: Install PyTorch with CUDA support if available:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118

